#18 refactoring test code
